backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic_core==2.18.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
psycopg2-binary==2.9.9
Pillow==10.3.0
bcrypt==3.2.0
